% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dr_clean_md_name}
\alias{dr_clean_md_name}
\title{Clean Dominican Republic municipal district names
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dr_clean_md_name(
  md,
  .mun = NULL,
  .prov = NULL,
  .tol = 0.33,
  .on_error = "fail"
)
}
\arguments{
\item{md}{Character vector of municipal district names to be cleaned.}

\item{.mun}{Optional character vector of municipality names. Defaults to NULL.
If provided, it will clean and match the municipal district names within the
context of these municipalities.}

\item{.prov}{Optional character vector of province names. Defaults to NULL.
If provided, it will clean and match the municipal district names within the
context of these provinces.}

\item{.tol}{Numeric tolerance level for string similarity. Defaults to 0.33.
This parameter controls how similar two strings must be to be considered a match.
A lower value means stricter matching.}

\item{.on_error}{Character string specifying the error handling method. Defaults to "fail".
It can be one of the following: "fail" to stop execution on error,
"omit" to ignore unmatched names, or "na" to return NA for unmatched names.}
}
\value{
A cleaned character vector of municipal district names.
}
\description{
This function cleans and standardizes the names of municipal districts (distritos municipales)
in the Dominican Republic. It can also handle related names for municipalities
and provinces, with tolerance for string similarity and options for error handling.
}
\examples{
\dontrun{
# Basic usage with municipal district names only
cleaned_md_names <- dr_clean_md_name(c("Distrito Nacion...", "Azua"))

# Usage with municipality names
cleaned_md_names <- dr_clean_md_name(
  c("DISTRITO NACIONAL", "Azua"),
  .mun = c("DISTRITO NACIONAL", "Azua de Compostela")
)
}

}
