% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\name{dr_clean_mun_name}
\alias{dr_clean_mun_name}
\title{Clean Dominican Republic municipality names
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dr_clean_mun_name(mun, .prov = NULL, .tol = 0.33, .on_error = "fail")
}
\arguments{
\item{mun}{Character vector of municipality names to be cleaned.}

\item{.prov}{Optional character vector of province names. Defaults to NULL.
If provided, it will clean and match the municipality names within the context
of these provinces.}

\item{.tol}{Numeric tolerance level for string similarity. Defaults to 0.33.
This parameter controls how similar two strings must be to be considered a match.
A lower value means stricter matching.}

\item{.on_error}{Character string specifying the error handling method. Defaults to "fail".
It can be one of the following: "fail" to stop execution on error,
"omit" to ignore unmatched names, or "na" to return NA for unmatched names.}
}
\value{
A cleaned character vector of municipality names.
}
\description{
This function cleans and standardizes the names of municipalities in the
Dominican Republic. It can also handle related names for provinces, with tolerance
for string similarity and options for error handling.
}
\examples{
\dontrun{
  # Basic usage with municipality names only
  cleaned_mun_names <- dr_clean_mun_name(c("Santo dominio este", "Azua de Compostela"))

  # Usage with province names
  cleaned_mun_names <- dr_clean_mun_name(
    c("SANTO DOMINGO DE GUZMAN", "Azua de Compostela"),
    .prov = c("Distrito Nacional", "Azua")
  )
}

}
