% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sections.R
\name{dr_clean_sec_name}
\alias{dr_clean_sec_name}
\title{Clean Dominican Republic section names
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dr_clean_sec_name(
  sec,
  .dm = NULL,
  .mun = NULL,
  .prov = NULL,
  .tol = 0.33,
  .on_error = "fail"
)
}
\arguments{
\item{sec}{Character vector of section names to be cleaned.}

\item{.dm}{Optional character vector of municipal district names. Defaults to NULL.
If provided, it will clean and match the section names within the context of
these municipal districts.}

\item{.mun}{Optional character vector of municipality names. Defaults to NULL.
If provided, it will clean and match the section names within the context of
these municipalities.}

\item{.prov}{Optional character vector of province names. Defaults to NULL.
If provided, it will clean and match the section names within the context of
these provinces.}

\item{.tol}{Numeric tolerance level for string similarity. Defaults to 0.33.
This parameter controls how similar two strings must be to be considered a match.
A lower value means stricter matching.}

\item{.on_error}{Character string specifying the error handling method.
Defaults to "fail". It can be one of the following: "fail" to stop execution on error,
"omit" to ignore unmatched names, or "na" to return NA for unmatched names.}
}
\value{
A cleaned character vector of section names.
}
\description{
This function cleans and standardizes the names of sections in the Dominican Republic.
It can also handle related names for municipal districts, municipalities,
and provinces, with tolerance for string similarity and options for error handling.
}
\examples{
\dontrun{
  # Basic usage with section names only
  cleaned_sec_names <- dr_clean_sec_name(
    c(
      "Santo Domingo de Guzmán (Zona urbana)",
      "Azua de Compost. (Zona urbana)"
     )
  )

  # Usage with municipal district names
  cleaned_sec_names <- dr_clean_sec_name(
    c(
      "SANTO DOMINGO DE GUZMAN (Zona urbana)",
      "Azua de Compostela (Zona urbana)"
    ),
    .dm = c("Distrito Nacional", "Azua")
  )
}

}
