% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R, R/municipalities.R, R/provinces.R,
%   R/regions.R, R/sections.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dr_municipal_districts}
\alias{dr_municipalities}
\alias{dr_provinces}
\alias{dr_regions}
\alias{dr_sections}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sfDR}{\code{\link[sfDR]{dr_municipal_districts}}, \code{\link[sfDR]{dr_municipalities}}, \code{\link[sfDR]{dr_provinces}}, \code{\link[sfDR]{dr_regions}}, \code{\link[sfDR]{dr_sections}}}
}}

