% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgisDR-package.R
\docType{package}
\name{rgisDR-package}
\alias{rgisDR}
\alias{rgisDR-package}
\title{rgisDR: GIS tools for Dominican Republic boundaries in R}
\description{
The rgisDR package provides simple features for the administrative divisions of the Dominican Republic, facilitating comprehensive geospatial analysis and mapping. It also includes tools for cleaning and standardizing administrative names, ensuring accurate and consistent naming conventions across provinces, municipalities, and municipal districts, among others.
}
\author{
\strong{Maintainer}: Daniel E. de la Rosa \email{me@dnldelarosa.com}

Other contributors:
\itemize{
  \item Adatar [copyright holder]
}

}
\keyword{internal}
